+++ HISTORY.md
@@ -4,6 +4,7 @@
 
 ### 2025-03-11
 * new dev version
+* extensions: new "tts_recognize" extension for voice-to-text
 
 ## 10.5.1:
 * **Notable Fixes:**